Fixed bug in actor spawn closures.
The closure in the block given to the actor_prim.spawn function could be
changed by the caller before the callee had a chance to extract its
values.

Fixed by resolving all pointers in the closure before the thread starts,
and storing them at a new glmalloced address. Changes to the pointers by
the caller are then no longer reflected in the closure.